# PERF: assert_produces_warning