--- a/pandas/_testing/_warnings.py
+++ b/pandas/_testing/_warnings.py
@@ -210,8 +210,8 @@ def _assert_raised_with_correct_stacklevel(
     # https://stackoverflow.com/questions/17407119/python-inspect-stack-is-slow
     frame = inspect.currentframe()
     for _ in range(4):
-        frame = frame.f_back
-    caller_filename = inspect.getfile(frame)
+        frame = frame.f_back  # type: ignore[union-attr]
+    caller_filename = inspect.getfile(frame)  # type: ignore[arg-type]
     msg = (
         "Warning not set with correct stacklevel. "
         f"File where warning is raised: {actual_warning.filename} != "