Similar to the implementation in `pandas.util._exceptions.find_stack_level`. 

Seems like it may save a little testing time:

```
> pytest pandas/tests/series/

160.85s  <- main
135.46s  <- PR
```

```
> pytest pandas/tests/arithmetic/

309.48s  <- main
276.90s  <- PR
```