@jbrockmendel 2023-11-17
also if there are a small number of highly-parametrized tests (seems likely in the arithmetic tests) that are responsible for a ton of these calls, can pass check_stacklevel=False for some/most of them